Añadida funcinoalidad de aprobar o rechazar voluntarios

+++ session/api.py
@@ -26,12 +26,15 @@ def get_sessions_by_organization(request):
 @api_view(['POST'])
 def insert_session(request):
     if request.user.is_authenticated:
+        if request.user.user_type == "Volunteer":
+            return Response({"message": "Un usuario de tipo voluntario no puede crear sesiones"}, status = status.HTTP_400_BAD_REQUEST)
         request.data["organization"] = request.user.id
         sessions_serializer = SessionSerializer(data = request.data)
         if sessions_serializer.is_valid():
             sessions_serializer.save()
             return Response(sessions_serializer.data, status = status.HTTP_200_OK)
         return Response(sessions_serializer.errors, status = status.HTTP_400_BAD_REQUEST)
+    
     return Response({"message": "Primero debe iniciar sesion"}, status = status.HTTP_400_BAD_REQUEST)
 
 
@@ -62,14 +65,50 @@ def insert_volunteer_in_session(request, id):
 
     if request.user.is_authenticated:
         volunteer = CustomUser.objects.filter(id = request.user.id).first()
-        if (volunteer == None):
+        if (volunteer == None or volunteer.user_type == "Organization"):
             return Response({"message": "El id ingresado no corresponde a un usuario validov"}, status = status.HTTP_400_BAD_REQUEST) 
         session = Session.objects.filter(id = id).first()
         session.volunteer.add(request.user.id)
         session.save()
-        sessions_serializer = SessionSerializer(session)
-        return Response(sessions_serializer.data, status = status.HTTP_200_OK)
+        request_volunteer = VolunteerRequest.objects.filter(session = session).filter(volunteer = volunteer).first()
+        if((request_volunteer == None)):
+            request_volunteer = VolunteerRequest.objects.create(status = 0,
+                                session = session, organization = session.organization, volunteer = volunteer)        
+            request_volunteer.save()
+            sessions_serializer = SessionSerializer(session)
+            return Response(sessions_serializer.data, status = status.HTTP_200_OK)
+        return Response({"message": "El Usuario ya se postulo para este voluntariado"}, status = status.HTTP_400_BAD_REQUEST) 
+
     return Response({"message": "Primero debe iniciar sesion"}, status = status.HTTP_400_BAD_REQUEST)
+
+@api_view(["PUT"])
+def approve_volunteer_in_session(request, idSession, idVolunteer):
+    if request.user.is_authenticated:
+        volunteer = CustomUser.objects.filter(id = idVolunteer).first()
+        if (volunteer == None or volunteer.user_type == "Organization"):
+            return Response({"message": "El id ingresado no corresponde a un usuario valido"}, status = status.HTTP_400_BAD_REQUEST) 
+        session = Session.objects.filter(id = idSession).first()
+        sessionVolunteer = VolunteerRequest.objects.filter(session = session).filter(volunteer = volunteer).first()
+        sessionVolunteer.status = 1
+        sessionVolunteer.save()
+        request_serializer = VolunteerRequestSerializer(sessionVolunteer)
+        return Response(request_serializer.data, status=status.HTTP_200_OK)
+    return Response({"message": "Primero debe iniciar sesion"}, status = status.HTTP_400_BAD_REQUEST)
+
+@api_view(["PUT"])
+def reject_volunteer_in_session(request, idSession, idVolunteer):
+    if request.user.is_authenticated:
+        volunteer = CustomUser.objects.filter(id = idVolunteer).first()
+        if (volunteer == None or volunteer.user_type == "Organization") :
+            return Response({"message": "El id ingresado no corresponde a un usuario valido"}, status = status.HTTP_400_BAD_REQUEST) 
+        session = Session.objects.filter(id = idSession).first()
+        sessionVolunteer = VolunteerRequest.objects.filter(session = session).filter(volunteer = volunteer).first()
+        sessionVolunteer.status = 2
+        sessionVolunteer.save()
+        request_serializer = VolunteerRequestSerializer(sessionVolunteer)
+        return Response(request_serializer.data, status=status.HTTP_200_OK)
+    return Response({"message": "Primero debe iniciar sesion"}, status = status.HTTP_400_BAD_REQUEST)
+
 
 
 
@@ -88,19 +127,26 @@ def update_session(request, id):
 
     if request.user.is_authenticated:    
         session = Session.objects.filter(id = id).first()
-        session_serializer = SessionSerializer(session, data = request.data)
-        if session_serializer.is_valid():
-            session_serializer.save()
-            return Response(session_serializer.data, status = status.HTTP_200_OK)
-        return Response(session_serializer.errors, status = status.HTTP_400_BAD_REQUEST)
+        if session.organization == request.user:
+            session_serializer = SessionSerializer(session, data = request.data)
+            if session_serializer.is_valid():
+                session_serializer.save()
+                return Response(session_serializer.data, status = status.HTTP_200_OK)
+            return Response(session_serializer.errors, status = status.HTTP_400_BAD_REQUEST)
+        return Response({"message" : "El voluntariado no pertenece a la organizacion que se encuentra actualmente autenticada"}
+                        , status = status.HTTP_400_BAD_REQUEST)
     return Response({"message": "Primero debe iniciar sesion"}, status = status.HTTP_400_BAD_REQUEST)
 
 
 @api_view(["DELETE"])
 def delete_session(request, id):
 
-    if request.user.is_authenticated:  
-        session = Session.objects.filter(id = id).first()
-        session.delete()
-        return Response({"message":"Sesion eliminada correctamente"}, status = status.HTTP_200_OK)
+    if request.user.is_authenticated: 
+        if session.organization == request.user: 
+            session = Session.objects.filter(id = id).first()
+            session.delete()
+            return Response({"message":"Sesion eliminada correctamente"}, status = status.HTTP_200_OK)
+        return Response({"message" : "El voluntariado no pertenece a la organizacion que se encuentra actualmente autenticada"}
+                        , status = status.HTTP_400_BAD_REQUEST)
+
     return Response({"message": "Primero debe iniciar sesion"}, status = status.HTTP_400_BAD_REQUEST)
